git: merge-index
